Add public interface for constructing and freeing caching allocators
